% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_request.R
\name{create_request}
\alias{create_request}
\title{Create request to EcoCounter API}
\usage{
create_request(EcoCounterId, path = c("data", "periode"),
  API_key = "cw6Xk4jW4X4R", query = NULL)
}
\arguments{
\item{EcoCounterId}{\link{character} or \link{integer} (\strong{requiered}): Id of the EcoCounter installation}

\item{path}{\link{character}: Path to requested data from EcoCounter API. Use c("data", "periode") or
c("counter", "channels").}

\item{API_key}{\link{character}: API key to get access to EcoCounter API}

\item{query}{\link{list}: Querys to get data for specific date. Use with argument path = c("data", "periode").}
}
\value{
\link{data.frame} with parsed data from EcoCounter API
}
\description{
Create request to EcoCounter API
}
\section{Function version 0.0.1}{
NA
}

\author{
Johannes Friedrich
}
