% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoCounteR-package.R
\docType{package}
\name{EcoCounteR-package}
\alias{EcoCounteR-package}
\title{EcoCounteR}
\description{
An R API wrapper for the Eco Counter project.
}
\details{
\tabular{ll}{Package: \tab EcoCounteR\cr Type: \tab Package\cr Version:
\tab 0.0.1 \cr Date: \tab 2018-XX-XX \cr License: \tab GPL-3\cr }
}
\keyword{package}
