% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EcoCounter_data.R
\name{get_EcoCounter_data}
\alias{get_EcoCounter_data}
\title{Get location from one EcoCounter installation}
\usage{
get_EcoCounter_data(EcoCounterId, from = NULL, to = NULL, step = 4)
}
\arguments{
\item{EcoCounterId}{\link{character} or \link{integer} (\strong{required}): Id of the EcoCounter installation}

\item{from}{\link{character}: Start date (YYYYMMDD). If no date is submitted, the day before yesterday is
automatically set.}

\item{to}{\link{character}: End date (YYYYMMDD). If no date is submitted, yesterday is
automatically set. Note: When date is the date of today, usually no data are available.}

\item{step}{\link{integer} or \link{character}: Get data : 2 (15 min), 3 (hourly), 4 (daily), 5 (weekly), 6 (monthly), 7 (yearly). With default: 4.}
}
\value{
\link{data.frame} with parsed data from EcoCounter API
}
\description{
Get location from one EcoCounter installation
}
\section{Function version 0.0.1}{
NA
}

\examples{

EcoCounterId <- 101020207
get_EcoCounter_data(EcoCounterId)

## set specific date and request weekly data (step = 5)
get_EcoCounter_data(EcoCounterId, from = "20171101", to = "20171201", step = 5)

}
\author{
Johannes Friedrich
}
