% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EcoCounter_info.R
\name{get_EcoCounter_info}
\alias{get_EcoCounter_info}
\title{Get specific information from one EcoCounter installation}
\usage{
get_EcoCounter_info(EcoCounterId)
}
\arguments{
\item{EcoCounterId}{\link{character} or \link{integer} (\strong{requiered}): Id of the EcoCounter installation}
}
\value{
\link{data.frame} with parsed data from EcoCounter API
}
\description{
Get specific information from one EcoCounter installation
}
\section{Function version 0.0.1}{
NA
}

\examples{

EcoCounterId <- 100020207
get_EcoCounter_info(EcoCounterId)

}
\author{
Johannes Friedrich
}
